fix: better typing to avoid warnings

+++ drvut_user/src/Connection.h
@@ -4,7 +4,7 @@
 #include "Buffer.h"
 
 template<typename T>
-concept Connection = requires(T& t, const Buffer& data, uint32_t size) {
+concept Connection = requires(T& t, const Buffer& data, size_t size) {
     { t.recv(size) } -> std::same_as<Buffer>;
-    { t.send(data) } -> std::same_as<uint32_t>;
+    { t.send(data) } -> std::same_as<size_t>;
 };
