+++ drvut_user/test/MockConnection.h
@@ -6,8 +6,8 @@
 
 class MockConnection {
 public:
-    MOCK_METHOD(Buffer, recv, (uint32_t));
-    MOCK_METHOD(uint32_t, send, (const Buffer&));
+    MOCK_METHOD(Buffer, recv, (size_t));
+    MOCK_METHOD(size_t, send, (const Buffer&));
 };
 
 class MoveableMockConnection {
@@ -15,11 +15,11 @@ public:
     MoveableMockConnection() 
         : m_mock(std::make_unique<testing::StrictMock<MockConnection>>()) {}
 
-    Buffer recv(uint32_t size) {
+    Buffer recv(size_t size) {
         return m_mock->recv(size);
     }
 
-    uint32_t send(const Buffer& data) {
+    size_t send(const Buffer& data) {
         return m_mock->send(data);
     }
